Refactor: resolve circular import
