+++ flexmeasures/api/v1_1/implementations.py
@@ -41,8 +41,8 @@ from flexmeasures.api.v1.implementations import (
 from flexmeasures.api.common.utils.api_utils import (
     get_sensor_by_generic_asset_type_and_location,
 )
-from flexmeasures.data.models.data_sources import get_or_create_source
 from flexmeasures.data.models.time_series import TimedBelief
+from flexmeasures.data.queries.data_sources import get_or_create_source
 from flexmeasures.data.services.resources import get_sensors
 from flexmeasures.data.services.forecasting import create_forecasting_jobs
