+++ flexmeasures/data/models/data_sources.py
@@ -1,10 +1,12 @@
-from typing import Optional, Union
+from __future__ import annotations
+from typing import Optional, TYPE_CHECKING
 
 import timely_beliefs as tb
-from flask import current_app
 
 from flexmeasures.data import db
-from flexmeasures.data.models.user import User, is_user
+
+if TYPE_CHECKING:
+    from flexmeasures.data.models.user import User
 
 
 class DataSource(db.Model, tb.BeliefSourceDBMixin):
@@ -84,42 +86,3 @@ class DataSource(db.Model, tb.BeliefSourceDBMixin):
 
     def __str__(self):
         return self.description
-
-
-def get_or_create_source(
-    source: Union[User, str],
-    source_type: Optional[str] = None,
-    model: Optional[str] = None,
-    flush: bool = True,
-) -> DataSource:
-    if is_user(source):
-        source_type = "user"
-    query = DataSource.query.filter(DataSource.type == source_type)
-    if model is not None:
-        query = query.filter(DataSource.model == model)
-    if is_user(source):
-        query = query.filter(DataSource.user == source)
-    elif isinstance(source, str):
-        query = query.filter(DataSource.name == source)
-    else:
-        raise TypeError("source should be of type User or str")
-    _source = query.one_or_none()
-    if not _source:
-        if is_user(source):
-            _source = DataSource(user=source, model=model)
-        else:
-            if source_type is None:
-                raise TypeError("Please specify a source type")
-            _source = DataSource(name=source, model=model, type=source_type)
-        current_app.logger.info(f"Setting up {_source} as new data source...")
-        db.session.add(_source)
-        if flush:
-            # assigns id so that we can reference the new object in the current db session
-            db.session.flush()
-    return _source
-
-
-def get_source_or_none(source: int, source_type: str) -> Optional[DataSource]:
-    query = DataSource.query.filter(DataSource.type == source_type)
-    query = query.filter(DataSource.id == int(source))
-    return query.one_or_none()
