+++ flexmeasures/data/models/parsing_utils.py
@@ -1,11 +1,9 @@
-from __future__ import annotations
-from typing import Optional, Union, List, TYPE_CHECKING
+from typing import Optional, Union, List
 
 from flask import current_app
 from flexmeasures.data import db
 
-if TYPE_CHECKING:
-    from flexmeasures.data.models.data_sources import DataSource
+from flexmeasures.data.models.data_sources import DataSource
 
 
 def parse_source_arg(
@@ -20,11 +18,11 @@ def parse_source_arg(
         sources = [source]
     else:
         sources = source
-    parsed_sources: List["DataSource"] = []
+    parsed_sources: List[DataSource] = []
     for source in sources:
         if isinstance(source, int):
             parsed_source = (
-                db.session.query("DataSource").filter_by(id=source).one_or_none()
+                db.session.query(DataSource).filter_by(id=source).one_or_none()
             )
             if parsed_source is None:
                 current_app.logger.warning(
@@ -33,9 +31,7 @@ def parse_source_arg(
             else:
                 parsed_sources.append(parsed_source)
         elif isinstance(source, str):
-            _parsed_sources = (
-                db.session.query("DataSource").filter_by(name=source).all()
-            )
+            _parsed_sources = db.session.query(DataSource).filter_by(name=source).all()
             if _parsed_sources is []:
                 current_app.logger.warning(
                     f"Beliefs searched for unknown source {source}"
